==> animal_farm/nutrient/STEP1_create_N_P_excretion_rasters.Rmd <==
---
title: "Excess nutrients from manure"
output: html_document
editor_options: 
  chunk_output_type: console
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(here)
library(stringr)
library(janitor)

source(here("_workflow/common.R"))
source(here("_workflow/common_spatial.R"))

un <- read_csv(here("_spatial/_output/UNSD_Methodology.csv")) %>% 
  dplyr::select(iso3c, Global_Name, Region_Name, Sub_region_Name, Intermediate_Region_Name)

n_exc <- read_csv(here("animal_farm/nutrient/data/excretion_rate_gf.csv")) %>% 
  select(-country)

n_to_p <- read_csv(here::here("animal_farm/nutrient/data/N_P_cf.csv")) %>% 
  select(animal, product, cf_gf) %>% 
  unique()

```



```{r}
files <- list.files(file.path(prep, "animal_farm/farm/location_tifs"), full.names = TRUE)

for(file in files){
  
#  file <- files[23]
    
  animal_name <- str_split(str_split(file, pattern = "/")[[1]][12], pattern = "_")[[1]][1]
  system_name <- str_split(str_split(file, pattern = "/")[[1]][12], pattern = "_")[[1]][2]
  product_name <- str_split(str_split(file, pattern = "/")[[1]][12], pattern = "_")[[1]][3]
  product_name <- str_remove(product_name, ".tif")
  
  ## Nitrogen excretion rasters
    n_exc_rast <- n_exc %>%
      filter(animal == animal_name & product == product_name) %>% 
      right_join(food_rgns_xy) %>%
      select(x,y, tonnes_N_head) %>% 
      rasterFromXYZ(crs = food_crs)
  
  rast <- raster(file)
  
  n_raster <- rast*n_exc_rast
  
  ## Phosphorous excretion rasters
  n_to_p_cf <- n_to_p %>%
    filter(animal == animal_name & product == product_name)  %>% 
    pull(cf_gf)

  p_raster <- calc(n_raster, fun=function(x){x*n_to_p_cf})
  
 # plot(log(1+p_raster))
  
  raster::writeRaster(n_raster,
                      file.path(paste(prep, "animal_farm/nutrient/N_from_excretion/", animal_name, "_", system_name, "_", product_name, "_N_excretion.tif", sep = "")),
                      format = "GTiff",
                      overwrite = TRUE)
    
  raster::writeRaster(p_raster,
                      file.path(paste(prep, "animal_farm/nutrient/P_from_excretion/", animal_name, "_", system_name, "_", product_name, "_P_excretion.tif", sep = "")),
                      format = "GTiff",
                      overwrite = TRUE)
  
}

```


Spot check some rasters

```{r}

## couple Ns

N_1 <- raster(file.path(paste(prep,"animal_farm/nutrient/N_from_excretion/cows_grassland_meat_N_excretion.tif", sep = "")))
N_2 <- raster(file.path(paste(prep,"animal_farm/nutrient/N_from_excretion/chickens_industrial_meat_N_excretion.tif", sep = "")))
N_3 <- raster(file.path(paste(prep,"animal_farm/nutrient/N_from_excretion/sheep_mixed_milk_N_excretion.tif", sep = "")))

## couple Ps
P_1 <- raster(file.path(paste(prep,"animal_farm/nutrient/P_from_excretion/cows_grassland_meat_P_excretion.tif", sep = "")))
P_2 <- raster(file.path(paste(prep,"animal_farm/nutrient/P_from_excretion/chickens_industrial_meat_P_excretion.tif", sep = "")))
P_3 <- raster(file.path(paste(prep, "animal_farm/nutrient/P_from_excretion/sheep_mixed_milk_P_excretion.tif", sep = "")))

par(mar=c(1,1,1,1))
plot(log(1+N_1))
plot(log(1+N_2))
plot(log(1+N_3))


plot(log(1+P_1))
plot(log(1+P_2))
plot(log(1+P_3))



```
