==> animal_farm/nutrient/STEP2_manure_P_leaching_rasters.Rmd <==
---
title: "Excess P leaching from manure"
output: html_document
editor_options: 
  chunk_output_type: console
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(here)
library(stringr)
library(doParallel)
library(janitor)
library(countrycode)

source(here("_workflow/common.R"))
source(here("_workflow/common_spatial.R"))

```

Calculate a P leaching rasters

```{r}
files <- list.files(file.path(prep, "animal_farm/nutrient/P_from_excretion/"), full.names = TRUE)

for(file in files){
  
  #file <- files[1]
  animal_name <- str_split(str_split(file, pattern = "/")[[1]][13], pattern = "_")[[1]][1]
  system_name <- str_split(str_split(file, pattern = "/")[[1]][13], pattern = "_")[[1]][2]
  product_name <- str_split(str_split(file, pattern = "/")[[1]][13], pattern = "_")[[1]][3]
  product_name <- str_remove(product_name, ".tif")
  
  rast <- raster(file)*0.065
  # par(mar=c(1,1,1,1))
  # plot(rast)

  raster::writeRaster(rast,
                      file.path(paste(prep, "animal_farm/nutrient/STEP2_P_leach/", animal_name, "_", system_name, "_", product_name, ".tif",  sep = "")),
                      format = "GTiff",
                      overwrite = TRUE)
}

```

Check 

```{r}

files <- list.files(file.path(prep, "animal_farm/nutrient/STEP2_P_leach"), full.names = TRUE)

total_p_leach_stack <- raster::stack(lapply(files,
                        raster))

total_p_leach_raster<- sum(total_p_leach_stack, na.rm = TRUE)

plot(log(1+total_p_leach_raster))

  country_sum <- zonal(total_p_leach_raster, food_rgns_tif, fun="sum", progress="text", na.rm=TRUE) %>% 
    data.frame() %>%
    rename(ID_0 = zone) %>%
    left_join(food_rgns, by="ID_0")

global_sum <- country_sum %>% 
  dplyr::summarise(total = sum(sum))

# 2,650,912 tonnes P 

```

"This simple calculation indicates that about 10% of the P inputs to soil are lost to aquatic systems. In the period 1970–2000 this represented 2–3 Tg P a−1"

This is 3,000,000 tonnes which is close.

https://agupubs.onlinelibrary.wiley.com/doi/full/10.1029/2009GB003576
